add/update: useFlippedのカスタムフック作成と該当箇所の差し替え
