+++ src/customHooks/index.ts
@@ -1,2 +1,3 @@
 export { default as useTyping } from './useTyping'
 export { default as useSetParams } from './useSetParams'
+export { default as useFlipped } from './useFlipped'
